Assistant/requirements.txt
# Core Framework
fastapi>=0.115.0  # Updated to support anyio>=4.5 for mcp compatibility
uvicorn[standard]>=0.30.0  # Updated to support anyio>=4.5
pydantic>=2.12.0  # Updated for openai>=2.6.1 compatibility
pydantic-settings>=2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1

# LLM Integration
openai>=2.6.1  # Supports both OpenAI and Azure OpenAI (v2.x)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20
langsmith>=0.0.80

# Vector Store & Memory
pgvector==0.2.4
chromadb==0.4.18

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-jaeger==1.21.0

# Caching & State
redis==5.0.1
hiredis==2.2.3

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pandas==2.1.3
numpy==1.26.2
pyyaml==6.0.1
httpx>=0.27.1,<1.0.0  # Updated to support mcp (requires >=0.27.1) while satisfying openai (<1.0.0)
tenacity==8.2.3

# MCP (Model Context Protocol)
mcp>=1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.11.0
ruff==0.1.6
mypy==1.7.1

